Puttering around with various line editing libs
