+++ impls/cpp.2/Makefile
@@ -1,6 +1,10 @@
-CC=g++
+CC=/bin/g++
 FLAGS=-g -Wall -Werror -Wextra
-INC=-Iincludes/.
+INC=-Iincludes/. -I
+
 
 step0_repl: step0_repl.cpp
 	$(CC) $(FLAGS) $(INC) step0_repl.cpp -o step0_repl
+
+step1_read_print: step1_read_print.cpp
+	$(CC) $(FLAGS) $(INC) step1_read_print.cpp -o step1_read_print
